+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The entire project (i.e., the GraphQL API, REST API, and server) use the same version number, and follows [semver](https://semver.org/). Backwards compatibility gets based off of the API (i.e., the GraphQL and REST APIs). For example, if the server has a backward incompatible change such as a DB schema update, but the APIs haven't changed, then the major version doesn't get bumped. Another example is if the GraphQL API hasn't changed, but the format of the HTTP request used to send the GraphQL document has changed, then the major version gets bumped.
 
-## v0.7.2
+## v0.8.0
 
 ### Server
 
@@ -13,8 +13,11 @@ The entire project (i.e., the GraphQL API, REST API, and server) use the same ve
 
 ### GraphQL
 
+- Removed:
+    - `Query.requestOnetimeToken`
 - Bug fixes:
     - `Mutation.emailEmailAddressVerification` now returns an error when supplied a verified email address.
+    - `Subscription`s used to have the access token sent in the URL, which was insecure. Now, they're sent after the connection opens as a text event.
 
 ### REST API